refactor: separate concerns between facade and service layers in articleLike, commentLike

+++ src/main/java/com/servertech/myboard/article/application/service/ArticleCommandService.java
@@ -4,6 +4,8 @@ import com.servertech.myboard.article.application.dto.request.CreateArticleReque
 import com.servertech.myboard.article.application.dto.request.UpdateArticleRequest;
 import com.servertech.myboard.article.domain.Article;
 import com.servertech.myboard.article.domain.ArticleRepository;
+import com.servertech.myboard.global.exception.EntityNotFoundException;
+import com.servertech.myboard.global.exception.UnauthorizedException;
 import com.servertech.myboard.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,22 +13,30 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class ArticleCommandService {
 	private final ArticleRepository articleRepository;
 
+	@Transactional
 	public Article save(CreateArticleRequest request, User user) {
 		Article article = Article.create(request.title(), request.content(), user);
 		return articleRepository.save(article);
 	}
 
-	public void deleteArticle(Long id) {
-		Article article = articleRepository.find(id).orElseThrow(() -> new IllegalArgumentException("Article not found"));
+	@Transactional
+	public void deleteArticle(Long id, User user) {
+		Article article = articleRepository.find(id).orElseThrow(() -> new EntityNotFoundException("Article not found"));
+		if (!article.getUser().getId().equals(user.getId())) {
+			throw new UnauthorizedException("You do not have permission to delete this article");
+		}
 		articleRepository.delete(article);
 	}
 
-	public void updateArticle(Long id, UpdateArticleRequest request) {
-		Article article = articleRepository.find(id).orElseThrow(() -> new IllegalArgumentException("Article not found"));
+	@Transactional
+	public void updateArticle(Long id, UpdateArticleRequest request, User user) {
+		Article article = articleRepository.find(id).orElseThrow(() -> new EntityNotFoundException("Article not found"));
+		if (!article.getUser().getId().equals(user.getId())) {
+			throw new UnauthorizedException("You do not have permission to update this article");
+		}
 		article.update(request.title(), request.content());
 	}
 }
